html/semantics/scripting-1/the-script-element/scriptexecute-events/009.html

<!DOCTYPE html>
<html>
  <title>Removing script in beforescriptexecute</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    targets = [];
    var t = async_test();
    document.addEventListener("beforescriptexecute",
      function(e) {
       targets.push(e.target);
       e.target.parentNode.removeChild(e.target);
    }, false);
 </script>
 <script>t.step(function(){assert_array_equals(targets, [document.getElementsByTagName("script")[3]])}); t.done();</script>
